Implement router for campgrounds route
