+++ index.js
@@ -4,9 +4,10 @@ const methodOverride = require('method-override');
 const mongoose = require('mongoose');
 const path = require('path');
 const PORT = 3000;
-const { campgroundSchema, reviewSchema } = require('./schemas');
+const { reviewSchema } = require('./schemas');
 const Campground = require('./models/campground');
 const Review = require('./models/review');
+const campgroundRoutes = require('./routes/campgrounds');
 const catchAsync = require('./utils/catchAsync');
 const ExpressError = require('./utils/ExpressError');
 
@@ -24,16 +25,6 @@ app.set('views', path.join(__dirname, 'views'));
 app.use(express.urlencoded({ extended: true }));
 app.use(methodOverride('_method'));
 
-const validateCampground = (req, res, next) => {
-  const { error } = campgroundSchema.validate(req.body);
-  if (error) {
-    const message = error.details.map(detail => detail.message).join(',');
-    throw new ExpressError(message, 400);
-  } else {
-    next();
-  }
-};
-
 const validateReview = (req, res, next) => {
   const { error } = reviewSchema.validate(req.body);
   if (error) {
@@ -44,68 +35,11 @@ const validateReview = (req, res, next) => {
   }
 };
 
+app.use('/campgrounds', campgroundRoutes);
+
 app.get('/', (req, res) => {
   res.render('home');
 });
-
-app.get(
-  '/campgrounds',
-  catchAsync(async (req, res) => {
-    const campgrounds = await Campground.find();
-    res.render('campgrounds/index', { campgrounds });
-  })
-);
-
-app.post(
-  '/campgrounds',
-  validateCampground,
-  catchAsync(async (req, res) => {
-    const campground = new Campground(req.body.campground);
-    await campground.save();
-    res.redirect(`/campgrounds/${campground._id}`);
-  })
-);
-
-app.get('/campgrounds/new', (req, res) => {
-  res.render('campgrounds/new');
-});
-
-app.get(
-  '/campgrounds/:id',
-  catchAsync(async (req, res) => {
-    const campground = await Campground.findById(req.params.id).populate(
-      'reviews'
-    );
-    res.render('campgrounds/show', { campground });
-  })
-);
-
-app.put(
-  '/campgrounds/:id',
-  validateCampground,
-  catchAsync(async (req, res) => {
-    const campground = await Campground.findByIdAndUpdate(req.params.id, {
-      ...req.body.campground,
-    });
-    res.redirect(`/campgrounds/${campground._id}`);
-  })
-);
-
-app.get(
-  '/campgrounds/:id/edit',
-  catchAsync(async (req, res) => {
-    const campground = await Campground.findById(req.params.id);
-    res.render('campgrounds/edit', { campground });
-  })
-);
-
-app.delete(
-  '/campgrounds/:id',
-  catchAsync(async (req, res) => {
-    await Campground.findByIdAndDelete(req.params.id);
-    res.redirect('/campgrounds');
-  })
-);
 
 app.post(
   '/campgrounds/:id/reviews',
